requirements.txt
flask
numpy
opencv-python-headless
pillow
matplotlib
gunicorn
